backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.9

# Validation
pydantic[email]==2.5.0

# Scheduling
apscheduler==3.10.4

# Email & Communication
aiosmtplib==3.0.1
requests==2.31.0
twilio==8.10.0

# File Processing
pillow==10.1.0
pypdf2==3.0.1

# PDF Generation
reportlab==4.0.7
qrcode==7.4.2

# Excel
openpyxl==3.1.2
pandas==2.1.4

# Utilities
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0